refactor: split ExplorerService into domain-focused sub-services
Break the monolithic ExplorerService (1,301 lines) into a BaseExplorerService
base class and 4 focused sub-services: PlaylistService, DetailService,
TasteService, and MemoryService. The slimmed ExplorerService retains only
analytics, browsing, and profile methods.

Move 3 hard-coded constants (playlist cache TTL, playlist fetch max,
enrichment cache TTL) into SettingsService.DEFAULTS so they are
admin-configurable at runtime via the settings API.

Closes #65
