+++ services/api/src/app/explorer/services/__init__.py
@@ -1,0 +1,13 @@
+"""Domain-focused explorer sub-services."""
+
+from app.explorer.services.detail_service import DetailService
+from app.explorer.services.memory_service import MemoryService
+from app.explorer.services.playlist_service import PlaylistService
+from app.explorer.services.taste_service import TasteService
+
+__all__ = [
+    "DetailService",
+    "MemoryService",
+    "PlaylistService",
+    "TasteService",
+]
